git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/branches/geant4/_symbols/09-05-ref-04_branch@58120 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ source/processes/hadronic/models/cascade/cascade/src/G4CascadeCoalescence.cc
@@ -404,7 +404,7 @@ void G4CascadeCoalescence::reportArgs(const G4String& name,
 				      const ClusterCandidate& aCluster) const {
   G4cout << " >>> G4CascadeCoalescence::" << name << " ";
   std::copy(aCluster.begin(), aCluster.end(),
-	    std::ostream_iterator<int>(G4cout, " "));
+	    std::ostream_iterator<size_t>(G4cout, " "));
   G4cout << G4endl;
 
   for (size_t i=0; i<aCluster.size(); i++)
